hot_fix_preprocessor

+++ Synthetic_tasks/Preprocessor/task1.cpp
@@ -4,8 +4,8 @@
 
  
 int main() {
-    unsigned long long a = -45;
-    unsigned long long b = 13;
-    unsigned long long c = -1;
+    unsigned a = -45;
+    unsigned b = 13;
+    unsigned c = -1;
     std::cout << (b * c) + a; // -58
 }
